+++ internal/handler/job/tags/add.go
@@ -10,8 +10,12 @@ package tags
 
 import (
 	"context"
+	"time"
 
 	"github.com/Southclaws/fault"
+	"github.com/go-openapi/strfmt"
+	"github.com/siemens/wfx/generated/model"
+	"github.com/siemens/wfx/internal/handler/job/events"
 	"github.com/siemens/wfx/middleware/logging"
 	"github.com/siemens/wfx/persistence"
 )
@@ -31,6 +35,17 @@ func Add(ctx context.Context, storage persistence.Storage, jobID string, tags []
 		contextLogger.Err(err).Msg("Failed to add tags to job")
 		return nil, fault.Wrap(err)
 	}
+
+	_ = events.PublishEvent(ctx, &events.JobEvent{
+		Ctime:  strfmt.DateTime(time.Now()),
+		Action: events.ActionAddTags,
+		Job: &model.Job{
+			ID:       updatedJob.ID,
+			ClientID: updatedJob.ClientID,
+			Workflow: updatedJob.Workflow,
+			Tags:     updatedJob.Tags,
+		},
+	})
 
 	contextLogger.Info().Msg("Added job tags")
 	return updatedJob.Tags, nil