+++ middleware/logging/writer.go
@@ -43,3 +43,10 @@ func (w *responseWriter) Write(b []byte) (int, error) {
 	n, err := w.bodyWriter.Write(b)
 	return n, fault.Wrap(err)
 }
+
+// Flush implements the http.Flusher interface.
+// This is used by the server-sent events implementation to flush a single event to the client.
+func (w *responseWriter) Flush() {
+	flusher := w.httpWriter.(http.Flusher)
+	flusher.Flush()
+}