+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add optional `description` field to workflows
+- Job event notifications via server-sent events (see #11)
 
 ### Fixed
 
@@ -21,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Refactored `wfxctl workflow delete` command to accept workflows as arguments instead of positional parameters
 - Prefer cgroup CPU quota over host CPU count
+- Empty or `null` arrays are omitted from JSON responses
 
 ### Removed